api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.62.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.6
distro==1.9.0
docker==7.1.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.100.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==6.0.0
MarkupSafe==3.0.2
msgspec==0.19.0
multidict==6.6.3
numpy==2.3.1
oauthlib==3.3.1
openai==1.99.9
packaging==25.0
pandas==2.3.0
Pillow==10.1.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-pptx==0.6.23
pytz==2025.2
reportlab==4.0.7
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
xlsxwriter==3.2.5
yarl==1.20.1
